refactor: ImageFile 엔티티 common 패키지로 이동
